Updates to wofparser tool
Bug fixes and add option to account for unpacked structure on older drivers

Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/68681
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>

+++ src/tools/ffdcparser/parser_common.h
@@ -246,7 +246,7 @@ uint64_t get_uint64(FILE* i_fhndl)
         byte = fgetc(i_fhndl);
         if(EOF != byte)
         {
-            ret |= (byte << (i*8));
+            ret |= ((uint64_t)byte << (i*8));
         }
     }
